added information on a pre-configured mykrobe image
added information on a pre-configured mykrobe image in which I setup the typhi panel, just to make it easy for people to run it, rather than having to download the mykrobe image, modify it locally, save it, and then access the edited container again (which might be confusing for people who aren't familiar with docker. as discussed with Kat.